Move http related classes into a Http namespace
